+++ host-software/keyplus/layout/ekc_data.py
@@ -26,6 +26,140 @@ class EKCData(object):
     # def to_bytes(self):
     #     return bytearray(self.data)
 
+    # def parse_json(self):
+    #     pass
+
+
+class EKCMouseGestureKey(EKCData):
+    # Data: {
+    #    0x00: KC_MOUSE_GESTURE
+    #    0x02: threshold
+    #    0x04: threshold_diag
+    #    0x06: threshold_tap
+    #    0x08: keycode left
+    #    0x0A: keycode right
+    #    0x0C: keycode up
+    #    0x0E: keycode down
+    #    0x10: keycode up_left
+    #    0x12: keycode up_right
+    #    0x14: keycode down_left
+    #    0x16: keycode down_right
+    #    0x18: keycode tap
+    # }
+    SIZE =  13*2
+
+    # Threshold for Horizontal and Vertical mouse gestures
+    GESTURE_THRESHOLD = 110
+
+    # Threshold for Diagonal mouse gestures
+    GESTURE_THRESHOLD_DIAG = 60
+
+    # Threshold for tap mouse gesture.  The mouse has to move less than this
+    # amount in both the X and Y axes to trigger a tap gesture.
+    GESTURE_THRESHOLD_TAP = 20
+
+    def __init__(self):
+        self.threshold = EKCMouseGestureKey.GESTURE_THRESHOLD
+        self.threshold_diag = EKCMouseGestureKey.GESTURE_THRESHOLD_DIAG
+        self.threshold_tap = EKCMouseGestureKey.GESTURE_THRESHOLD_TAP
+
+        self.kc_left = None
+        self.kc_right = None
+        self.kc_up = None
+        self.kc_down = None
+
+        self.kc_up_left = None
+        self.kc_up_right = None
+        self.kc_down_left = None
+        self.kc_down_right = None
+
+        self.kc_tap = None
+
+    def size(self):
+        return self.SIZE
+
+    def to_bytes(self):
+        result = bytearray(self.SIZE)
+
+        struct.pack_into("< 13H", result, 0,
+            keycodes.KC_MOUSE_GESTURE,
+            self.threshold,
+            self.threshold_diag,
+            self.threshold_tap,
+            self.kc_map_function(self.kc_left),
+            self.kc_map_function(self.kc_right),
+            self.kc_map_function(self.kc_up),
+            self.kc_map_function(self.kc_down),
+            self.kc_map_function(self.kc_up_left),
+            self.kc_map_function(self.kc_up_right),
+            self.kc_map_function(self.kc_down_left),
+            self.kc_map_function(self.kc_down_right),
+            self.kc_map_function(self.kc_tap),
+        )
+
+        return result
+
+    def parse_json(self, kc_name, json_obj=None, parser_info=None):
+
+        print_warnings = False
+
+        if parser_info == None:
+            assert(json_obj != None)
+            print_warnings = True
+            parser_info = KeyplusParserInfo(
+                "<EKCMouseGestureKey Dict>",
+                {kc_name : json_obj}
+            )
+
+        parser_info.enter(kc_name)
+
+        # Get the tap key field
+        self.keycode = parser_info.try_get(
+            'keycode',
+            field_type=str
+        )
+        assert_equal(self.keycode, 'mouse_gesture')
+
+        # Get movement thresholds for gestures
+        self.threshold = parser_info.try_get(
+            'threshold',
+            field_type=int,
+            default=EKCMouseGestureKey.GESTURE_THRESHOLD
+        )
+        self.threshold_diag = parser_info.try_get(
+            'threshold_diag',
+            field_type=int,
+            default=EKCMouseGestureKey.GESTURE_THRESHOLD_DIAG
+        )
+        self.threshold_tap = parser_info.try_get(
+            'threshold_tap',
+            field_type=int,
+            default=EKCMouseGestureKey.GESTURE_THRESHOLD_TAP
+        )
+
+        # Get the keycodes for different directions field
+        self.kc_left  = parser_info.try_get('left' , default='none')
+        self.kc_right = parser_info.try_get('right', default='none')
+        self.kc_up    = parser_info.try_get('up'   , default='none')
+        self.kc_down  = parser_info.try_get('down' , default='none')
+
+        self.kc_up_left    = parser_info.try_get('up_left'   , default='none')
+        self.kc_up_right   = parser_info.try_get('up_right'  , default='none')
+        self.kc_down_left  = parser_info.try_get('down_left' , default='none')
+        self.kc_down_right = parser_info.try_get('down_right', default='none')
+
+        self.kc_tap = parser_info.try_get('tap', default='none')
+
+        # Finish parsing `device_name`
+        parser_info.exit()
+
+        # If this is debug code, print the warnings
+        if print_warnings:
+            for warn in parser_info.warnings:
+                print(warn, file=sys.stderr)
+
+
+
 class EKCHoldKey(EKCData):
     # Layout option:
     #   1: kc_hold_key
@@ -57,7 +191,7 @@ class EKCHoldKey(EKCData):
         self.activate_type = activate_type or EKCHoldKey.DEFAULT_ACTIVATE_TYPE
 
     def size(self):
-        return EKCHoldKey.SIZE
+        return self.SIZE
 
     def check_activate_type_valid(self, type_):
         if type_ not in EKCHoldKey.ACTIVATE_TYPE_MAP:
@@ -70,7 +204,7 @@ class EKCHoldKey(EKCData):
             )
 
     def to_bytes(self):
-        result = bytearray(EKCHoldKey.SIZE)
+        result = bytearray(self.SIZE)
 
         option_data = 0
 
@@ -115,16 +249,10 @@ class EKCHoldKey(EKCData):
         assert_equal(self.keycode, 'hold')
 
         # Get the hold key field
-        self.hold_key = parser_info.try_get(
-            'hold_key',
-            field_type=str
-        )
+        self.hold_key = parser_info.try_get('hold_key')
 
         # Get the tap key field
-        self.tap_key = parser_info.try_get(
-            'tap_key',
-            field_type=str
-        )
+        self.tap_key = parser_info.try_get('tap_key')
 
         # Get the delay key field
         self.delay = parser_info.try_get(
@@ -198,6 +326,11 @@ class EKCDataTable(EKCData):
 
         return result
     pack = to_bytes
+
+EKCKeycodeTable = {
+    'hold': EKCHoldKey,
+    'mouse_gesture': EKCMouseGestureKey,
+}
 
 if __name__ == '__main__':
     ekc_data = EKCDataTable()